requirements.txt
strawberry-graphql[debug-server] >= 0.57.4
dataclasses-json
chalice